requirements.txt
pillow==11.0.0
pillow-avif-plugin==1.4.6
pillow_heif==0.20.0
# notebook==7.2.2